#133 [REFACTOR] 알람 도메인 컨트롤러 패키지 생성 및 이전
